pfp component, browser util
