+++ src/utils/index.ts
@@ -1,6 +1,8 @@
 import type { NextRouter } from "next/router";
 import type { Query } from "nextjs-routes";
 
+export const browser = typeof window !== "undefined";
+
 export function getQueryParam(query: Query[string]) {
   if (query === undefined) return undefined;
   return Array.isArray(query) ? query.at(-1) : query;
